requirements_test.txt
pytest==8.*
pytest-asyncio==0.24.*
pytest-homeassistant-custom-component==0.13.*
ruff==0.8.*
